+++ src/test/java/io/knotx/commons/http/request/AllowedHeadersFilterTest.java
@@ -15,6 +15,7 @@
  */
 package io.knotx.commons.http.request;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
@@ -30,31 +31,42 @@ class AllowedHeadersFilterTest {
     private static final String CONTENT_TYPE_LOWERCASE = "content-type";
     private static final String ACCEPT = "Accept";
     private static final String ACCEPT_LOWERCASE = "accept";
+    private static final String UPGRADE_INSECURE_REQUESTS = "Upgrade-Insecure-Requests";
+    private static final String WILDCARD = ".*";
     private static final String[] HEADERS = {
             ACCEPT,
             "Accept-Encoding",
             "Accept-Language",
             "Connection",
             "Host",
-            "Upgrade-Insecure-Requests",
+            UPGRADE_INSECURE_REQUESTS,
             "User-Agent",
             "Keep-Alive"
     };
+    private static final String[] WILDCARD_ALLOWED_HEADERS = {
+        ACCEPT,
+        "Accept-.*",
+        "Connection",
+        "Host",
+        "Upgrade-Insecure-.*",
+        "User-.*",
+        "Keep-.*"
+    };
 
-    private static List<Pattern> preparePatters() {
-        return Arrays.asList(HEADERS).stream()
+    private static List<Pattern> preparePatters(String [] allowedHeaders) {
+        return Arrays.stream(allowedHeaders)
             .map(Pattern::compile)
             .collect(Collectors.toList());
     }
 
-    public static AllowedHeadersFilter getPatternFilter() {
-        return AllowedHeadersFilter.create(preparePatters());
+    public static AllowedHeadersFilter getPatternFilter(String [] allowedHeaders) {
+        return AllowedHeadersFilter.create(preparePatters(allowedHeaders));
     }
 
     @Test
     @DisplayName("Check header with allowed headers list")
     void checkHeaderWithAllowedHeadersList() {
-        AllowedHeadersFilter allowedHeadersFilter = getPatternFilter();
+        AllowedHeadersFilter allowedHeadersFilter = getPatternFilter(HEADERS);
         assertFalse(allowedHeadersFilter.test(CONTENT_TYPE));
         assertTrue(allowedHeadersFilter.test(ACCEPT));
     }
@@ -62,7 +74,7 @@ class AllowedHeadersFilterTest {
     @Test
     @DisplayName("Check header with allowed headers list as case sensitive")
     void checkHeaderWithAllowedHeadersListCaseSensitive() {
-        AllowedHeadersFilter allowedHeadersFilter = getPatternFilter();
+        AllowedHeadersFilter allowedHeadersFilter = getPatternFilter(HEADERS);
         assertFalse(allowedHeadersFilter.test(ACCEPT_LOWERCASE));
         assertTrue(allowedHeadersFilter.test(ACCEPT));
     }
@@ -99,5 +111,29 @@ class AllowedHeadersFilterTest {
         Set<String> strings = new HashSet<>();
         AllowedHeadersFilter allowedHeadersFilter = AllowedHeadersFilter.CaseInsensitive.create(strings);
         assertFalse(allowedHeadersFilter.test(CONTENT_TYPE));
+    }
+
+    @Test
+    @DisplayName("Check header with wildcards variants allowed headers list")
+    void checkHeaderWithWildcardsVariantsAllowedHeadersList() {
+        Set<String> allowedHeaders = new HashSet<>(Arrays.asList(WILDCARD_ALLOWED_HEADERS));
+        List<String> headers = Arrays.asList(HEADERS);
+
+        List<String> filteredHeaders = headers.stream()
+            .filter(AllowedHeadersFilter.CaseInsensitive.create(allowedHeaders))
+            .collect(Collectors.toList());
+        assertEquals(filteredHeaders, headers);
+    }
+
+    @Test
+    @DisplayName("Check header with single wildcard sign")
+    void checkHeaderWithSingleWildcardSignAllowedHeadersList() {
+        Set<String> allowedHeaders = new HashSet<>(Collections.singletonList(WILDCARD));
+        List<String> headers = Arrays.asList(HEADERS);
+
+        List<String> filteredHeaders = headers.stream()
+            .filter(AllowedHeadersFilter.CaseInsensitive.create(allowedHeaders))
+            .collect(Collectors.toList());
+        assertEquals(filteredHeaders, headers);
     }
 }