+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to `knotx-commons` will be documented in this file.
 
 ## Unreleased
 List of changes that are finished but not yet released in any final version.
+- [PR-14](https://github.com/Knotx/knotx-commons/pull/14) - KnotxServer response configuration - wildcards [41](https://github.com/Knotx/knotx-server-http/issues/41)
 - [PR-8](https://github.com/Knotx/knotx-commons/pull/8) - Unit Test for HTTP request commons
 
 ## 2.1.0